[RENAME] LeagueResponseForManager 를 LeagueResponseWithInProgressGames 로 변경 #192
